video_engine/requirements.txt
moviepy==1.0.3
google-cloud-texttospeech==2.21.0
google-api-python-client==2.151.0
boto3==1.35.49

